css hiding/show instead of injecting/removing
These changes give an option for event driven applications to use flippy
without worrying about their content from losing event bindings when DOM
elements are removed. Instead of using inject/remove, setting noRemove
will watch for a given selector. If the selector is already in the DOM,
it will make it visible when the flipbox is flipped.